azure-cosmos-mcp-server-samples/python/requirements.txt
azure-cosmos
pandas
mcp[cli]
fastmcp
azure-identity
python-dotenv